backend/bensco/backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers==4.3.1
gunicorn
dj-database-url
psycopg2-binary
whitenoise

